+++ robotframework-ls/tests/robotframework_ls_tests/completions/test_auto_import.py
@@ -1,19 +1,9 @@
-import pytest
-from typing import Dict, List
 from pathlib import Path
 import time
+from typing import Dict, List
 
-
-def apply_completion(doc, completion, expect_additional_text_edits=True):
-    text_edit = completion["textEdit"]
-    additional_text_edits = completion["additionalTextEdits"]
-    if expect_additional_text_edits:
-        assert additional_text_edits
-    else:
-        assert not additional_text_edits
-    doc.apply_text_edits([text_edit])
-    if additional_text_edits:
-        doc.apply_text_edits(additional_text_edits)
+import pytest
+from .common import apply_completion
 
 
 def test_completion_with_auto_import_basic_stdlib(